pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pproc"
description = "ECMWF Post-processing tools"
version = "1.4.0"
readme = "README.md"
license = { text = "Apache License Version 2.0" }
authors = [
    { name = "European Centre for Medium-Range Weather Forecasts (ECMWF)", email = "software.support@ecmwf.int" },
]

requires-python = ">=3.8"
dependencies = [
    "code-meters >= 0.0.3",
    "earthkit-data >= 0.13.2",
    "earthkit-meteo >= 0.1.1",
    "earthkit-time >= 0.1.3",
    "eccodes >= 1.6.1",
    "filelock >= 3.12.0",
    "mir-python >= 0.2.0",
    "numexpr",
    "numpy",
    "pandas >= 1.3.3",
    "psutil",
    "pyfdb",
    "thermofeel >= 2.1.0",
    "scikit-learn",
    "scipy >= 1.8",
    "xarray",
    "pydantic",
    "conflator >= 0.1.7",
]

[project.optional-dependencies]
tcycl = ["pyinfero >= 0.1.1", "wget"]
test = ["pytest", "requests"]

[project.scripts]
pproc-extreme-forecast-simple = "pproc.extreme_forecast_simple:main"
pproc-extreme = "pproc.extreme:main"
pproc-interpol = "pproc.interpol:main"
pproc-pts = "pproc.pts:main"
pproc-probabilities = "pproc.probabilities:main"
pproc-ensms = "pproc.ensms:main"
pproc-wind = "pproc.wind:main"
pproc-spectrum = "pproc.spectrum:main"
pproc-clustereps = "pproc.clustereps.__main__:main"
pproc-clustereps-pca = "pproc.clustereps.pca:main"
pproc-clustereps-cluster = "pproc.clustereps.cluster:main"
pproc-clustereps-attr = "pproc.clustereps.attribution:main"
pproc-anomaly-probabilities = "pproc.anomaly_probs:main"
pproc-quantiles = "pproc.quantiles:main"
pproc-tcycl-run = "pproc.tcycl.tcycl_run:main"
pproc-tcycl-summarise-ibtks = "pproc.tcycl.tcycl_summarise_ibtks:main"
pproc-tcycl-summarise-tcycl = "pproc.tcycl.tcycl_summarise_tcycl:main"
pproc-tcycl-summarise-trckr = "pproc.tcycl.tcycl_summarise_trckr:main"
pproc-tcycl-evaluate = "pproc.tcycl.tcycl_evaluate:main"
pproc-thermal-indices = "pproc.thermal_indices:main"
pproc-histogram = "pproc.histogram:main"
pproc-accumulate = "pproc.accumulate:main"
pproc-significance = "pproc.significance:main"
pproc-anomaly = "pproc.anomaly:main"
pproc-monthly-stats = "pproc.monthly_stats:main"
pproc-config="pproc.config_generation:main"

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests"]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
